+++ config/config.exs
@@ -2,4 +2,6 @@ import Config
 
 config :rock_solid, cache_enabled: false
 
+config :cuerdo, :screen, Cuerdo.CLI.Screen.Terminal
+
 import_config "#{config_env()}.exs"