+++ lib/cuerdo/cli.ex
@@ -7,10 +7,7 @@ defmodule Cuerdo.CLI do
   - `--exclude` - Comma-separated list of workflow ids to exclude from the document
   - `--only` - Comma-separated list of workflow ids to execute from the document
   - `--halt-on-error` (flag) - Whether to stop execution on the first failure of each workflow
-  - `--report-output` - The test suite report output. Defaults to `stdout`. Supported values
-  are `stdout` and `json`
-  - `--report-file` - The report file destination. Mandatory if `--report-output` is different
-  from `stdout`
+  - `--report-file` - The report file destination. Defaults to `report.json`
 
   For more information on the options, refer to `Cuerdo.ArazzoCase`
   """
@@ -26,7 +23,6 @@ defmodule Cuerdo.CLI do
   @impl true
   def start(_, _) do
     if Application.get_env(:cuerdo, :run_cli, true) do
-      Logger.info("Running Cuerdo Arazzo runner")
       main(Burrito.Util.Args.argv())
     end
 
@@ -36,14 +32,12 @@ defmodule Cuerdo.CLI do
   def main(args) do
     case run(args) do
       {:error, _} ->
-        Logger.flush()
         System.halt(1)
 
       {:ok, results} when is_list(results) ->
         status =
           if(Enum.reject(results, &(&1.status == :passed)) |> Enum.empty?(), do: 0, else: 1)
 
-        Logger.flush()
         System.halt(status)
     end
   end
@@ -51,25 +45,33 @@ defmodule Cuerdo.CLI do
   def run([]) do
     msg = "Empty arguments. Pass path/to/arazzo.yaml"
     Logger.error(msg)
+    Logger.flush()
     {:error, %ArgumentError{message: msg}}
   end
 
   def run([document_path | args]) do
+    CLI.Screen.start()
+
     with {:ok, valid_args} <- CLI.Args.parse(args),
          {:ok, document} <- YamlElixir.read_from_file(document_path),
+         _ <- CLI.Screen.fetched_document(),
          {:ok, parsed_doc} <- Arazzo.Document.new(document),
          {:ok, workflow_ids} <-
            ArazzoCase.Runner.workflow_ids(parsed_doc, valid_args[:only], valid_args[:exclude]) do
       opts = Keyword.put(valid_args, :document, document)
-      Logger.info("Executing workflows: #{Enum.join(workflow_ids, ", ")}")
+
+      CLI.Screen.start_workflows(workflow_ids, opts[:num_runs])
       results = Enum.flat_map(workflow_ids, &ArazzoCase.Runner.run_all(&1, document, opts))
 
-      ArazzoCase.Report.write(Keyword.fetch!(opts, :report_output), results, opts[:report_file])
+      report_file = Keyword.fetch!(opts, :report_file)
+      ArazzoCase.Report.write(:json, results, report_file)
+      CLI.Screen.summary(results, report_file)
 
       {:ok, results}
     else
       {:error, exc} = error when is_exception(exc) ->
         Logger.error("Error processing arguments/document: #{Exception.message(exc)}")
+        Logger.flush()
         error
     end
   end